Update 02-create-app.md
There is an additional change needed for newer versions of Laravel to provide the namespace for routing.  This error occurred for me with: Laravel Framework 8.14.0.

It was fixed by following the instruction from MichalOravec, here: https://laracasts.com/discuss/channels/laravel/laravel-8-error-target-class-homecontroller-does-not-exist

"To App/Providers/RouteServiceProvider.php add $namespace"

```
class RouteServiceProvider extends ServiceProvider
{
    /**
     * This namespace is applied to your controller routes.
     *
     * In addition, it is set as the URL generator's root namespace.
     *
     * @var string
     */
    protected $namespace = 'App\Http\Controllers';
}
```

The comment on the linked laracasts page also mentions making sure the boot method uses $namespace, but that was already in place for my version of Laravel, so I am unsure when that is needed.